Region of interest should contain floats, not ints

+++ remove_restore.py
@@ -56,7 +56,7 @@ def main():
     parser.add_argument('--diff_threshold', default=0.0, help='value above which differences will be added to the base grid')
     parser.add_argument('--plot', action='store_true', help='plot final output before saving')
     parser.add_argument('--output', required=True, help='filename of final output')
-    parser.add_argument('--region_of_interest', required=False, nargs=4, type=int,
+    parser.add_argument('--region_of_interest', required=False, nargs=4, type=float,
                         help='output region in order <xmin> <xmax> <ymin> <ymax>')
     args = parser.parse_args()
